Various files updated/removed for omptimisation and fuctionality
